Improve linking scripts
Linking scripts now supports machine specific scripts, preventing
scripts being linked on machines that don't need it.

+++ copy_scripts.sh
@@ -2,19 +2,23 @@
 
 . ./color.sh
 
-SCRIPTS_DIR="$(pwd)/.scripts/*"
+SCRIPTS_DIR="$(pwd)/.scripts"
 
 link_scripts() {
     # Link the general scripts
-
-    # First make the ~/.scripts folder if it doesn't exist
-    mkdir -p ~/.scripts
-
-    for FILE in $SCRIPTS_DIR; do
-        echo -e "${GREEN}Linking script: $FILE...${NOCOLOR}"
-        FILENAME=${FILE##*/}
-        ln -sf $FILE ~/.scripts/$FILENAME
-    done
+    
+    # Check for machine specific scripts
+    if [ -e $SCRIPTS_DIR/$HOSTNAME ]; then
+        # Make the directory if it doesn't exist
+        mkdir -p ~/.scripts
+        
+        # Link each machine specific script
+        for FILE in $SCRIPTS_DIR/$HOSTNAME/*; do
+            echo -e "${GREEN}Linking script: $FILE...${NOCOLOR}"
+            FILENAME=${FILE##*/}
+            ln -sf $FILE ~/.scripts/$FILENAME
+        done
+    fi
 }
 
 # Run the link
